Update language/include/vmgcdata.h - GCData structure - Use unsigned short for nReferenceCount

+++ extensions/webassembly/ringqt/project/ring/include/vmgcdata.h
@@ -2,7 +2,7 @@
 #ifndef ring_vmgcdata
     #define ring_vmgcdata
     typedef struct GCData {
-        int nReferenceCount  ;
+        unsigned short nReferenceCount  ;
         void (*pFreeFunc)(void *,void *) ;
     } GCData ;
 #endif
